fix -D

+++ src/panmanUtils.cpp
@@ -136,7 +136,7 @@ void setupOptionDescriptions() {
     ("input-msa,M", po::value< std::string >(), "Input MSA file (FASTA format) to build a PanMAN")
     ("input-newick,N", po::value< std::string >(), "Input tree topology as Newick string")
     ("impute", "Create new PanMAN with N sequences inputed")
-    ("max-insertion-impute-distance,D", po::value< int8_t >(), "Maximum total branch length to move a node for insertion imputation [default 5]")
+    ("max-insertion-impute-distance,D", po::value< int64_t >(), "Maximum total branch length to move a node for insertion imputation [default 5]")
     ("create-network,K",po::value< std::vector<std::string>>(), "Create PanMAN with network of trees from single or multiple PanMAN files")
 
     // ("optimize", "currently UNSUPPORTED: whether given msa file should be optimized or not")
@@ -188,7 +188,7 @@ void setupOptionDescriptions() {
 
     imputeDesc.add_options()
         ("input-file", po::value< std::string >(), "Input file name")
-        ("max-insertion-impute-distance,D", po::value< int8_t >(), 
+        ("max-insertion-impute-distance,D", po::value< int64_t >(), 
         "Maximum total branch length to move a node for insertion imputation [default 5]")
         ("output-file,o", po::value< std::string >(), "Output file name");
 
@@ -336,7 +336,7 @@ void impute(panmanUtils::TreeGroup *TG, po::variables_map &globalVm, std::ofstre
 
     int allowedDistance = 5;
     if(globalVm.count("max-insertion-impute-distance")) {
-        allowedDistance = globalVm["max-insertion-impute-distance"].as< int8_t >();
+        allowedDistance = globalVm["max-insertion-impute-distance"].as< int64_t >();
     }
 
     auto imputeStart = std::chrono::high_resolution_clock::now();
